change not-found to server component
